requirements.txt
openai
python-dotenv
truss
transformers
requests